requirements.txt
streamlit
openai
opencv-python
Pillow